requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
psycopg2-binary==2.9.9
pandas==2.2.3
PyMuPDF==1.24.9
nltk==3.9.1
sumy==0.11.0
lxml==5.3.0
tqdm==4.66.5
python-dotenv==1.0.1
